REL-4170: simple test case
