+++ bundle/edu.gemini.pot/src/test/scala/edu/gemini/spModel/template/InstantiationFunctorTest.scala
@@ -1,23 +1,13 @@
 package edu.gemini.spModel.template
 
-import edu.gemini.pot.sp.{Instrument, ISPObsComponent, SPComponentType, ISPObservation, SPNodeKey}
-import edu.gemini.pot.sp.SPComponentType.{INSTRUMENT_GMOSSOUTH, OBSERVER_OBSERVE, SCHEDULING_CONDITIONS, TELESCOPE_TARGETENV}
-import edu.gemini.shared.util.TimeValue
-import edu.gemini.spModel.data.ISPDataObject
+import edu.gemini.pot.sp.{Instrument, ISPObservation}
+import edu.gemini.pot.sp.SPComponentType.INSTRUMENT_GMOSSOUTH
 import edu.gemini.spModel.gemini.gmos.GmosSouthType.FilterSouth
 import edu.gemini.spModel.gemini.gmos.InstGmosSouth
 import edu.gemini.spModel.gemini.gmos.blueprint.SpGmosSBlueprintImaging
-import edu.gemini.spModel.gemini.obscomp.SPSiteQuality
-import edu.gemini.spModel.gemini.obscomp.SPSiteQuality.CloudCover
-import edu.gemini.spModel.gemini.obscomp.SPSiteQuality.CloudCover.{PERCENT_50 => CC50, PERCENT_80 => CC80}
 import edu.gemini.spModel.gemini.security.UserRolePrivileges
-import edu.gemini.spModel.obsclass.ObsClass
 import edu.gemini.spModel.obsclass.ObsClass._
-import edu.gemini.spModel.seqcomp.SeqRepeatObserve
-import edu.gemini.spModel.target.SPTarget
-import edu.gemini.spModel.target.env.TargetEnvironment
-import edu.gemini.spModel.target.obsComp.TargetObsComp
-import edu.gemini.spModel.test.SpModelTestBase
+
 import org.junit.Assert._
 import org.junit.Test
 
@@ -25,9 +15,8 @@ import scala.collection.JavaConverters._
 import scalaz._
 import Scalaz._
 
-import InstantiationFunctorTest._
 
-class InstantiationFunctorTest extends SpModelTestBase {
+class InstantiationFunctorTest extends TestBase {
 
   def instantiate(os: ISPObservation*): List[ISPObservation] = {
     val fact = getFactory
@@ -49,7 +38,7 @@ class InstantiationFunctorTest extends SpModelTestBase {
     val tgNode = fact.createTemplateGroup(prog, WithSomeNewKey)
     val tg     = new TemplateGroup()
     tg.setBlueprintId("blueprint-0")
-    tfNode.setDataObject(tg)
+    tgNode.setDataObject(tg)
 
     val tpNode = fact.createTemplateParameters(prog, WithSomeNewKey)
     val tp     = newParameters(ScienceTargetName, ScienceTargetCC)
@@ -85,48 +74,26 @@ class InstantiationFunctorTest extends SpModelTestBase {
     getProgram.getAllObservations.asScala.toList.filter(o => obsKeys(o.getNodeKey))
   }
 
-  def newObs(clazz: ObsClass): ISPObservation = {
-    val obsNode     = getFactory.createObservation(getProgram, Instrument.GmosSouth.some, WithSomeNewKey)
 
-    val filteredChildren = obsNode.getObsComponents.asScala.filter { n =>
-      n.getDataObject.getType match {
-        case TELESCOPE_TARGETENV | SCHEDULING_CONDITIONS => false
-        case _                                           => true
-      }
+  def getPositionAngle(o: ISPObservation): Double =
+    obsComp[InstGmosSouth](o, INSTRUMENT_GMOSSOUTH).map {
+      case (_, gmos) => gmos.getPosAngleDegrees
+    } | sys.error("expected gmos component")
+
+  def setPositionAngle(o: ISPObservation, pa: Double): Unit =
+    obsComp[InstGmosSouth](o, INSTRUMENT_GMOSSOUTH).foreach {
+      case (oc, gmos) => gmos.setPosAngleDegrees(pa)
+                         oc.setDataObject(gmos)
+      case _          => fail("expected gmos component")
     }
-    obsNode.setObsComponents(filteredChildren.asJava)
 
-    val observeNode = getFactory.createSeqComponent(getProgram, OBSERVER_OBSERVE, WithSomeNewKey)
-    val observe     = new SeqRepeatObserve() <| (_.setObsClass(clazz))
-    observeNode.setDataObject(observe)
-
-    obsNode.getSeqComponent.addSeqComponent(observeNode)
-    obsNode
-  }
-
-  def newObsComp(cType: SPComponentType, dob: ISPDataObject): ISPObsComponent =
-    getFactory.createObsComponent(getProgram, cType, WithSomeNewKey) <| (_.setDataObject(dob))
-
-  def newSiteQualityComponent(cc: CloudCover): ISPObsComponent =
-    newObsComp(SCHEDULING_CONDITIONS, newSiteQuality(cc))
-
-  def newTargetComponent(n: String): ISPObsComponent =
-    newObsComp(TELESCOPE_TARGETENV, newTargetObsComp(n))
-
-  def editObservation(o: ISPObservation): Unit = {
-    def editObsComponent(o: ISPObservation, ct: SPComponentType, noc: ISPObsComponent): Unit =
-      o.getObsComponents.asScala.find { _.getType == ct }.fold(o.addObsComponent(noc)) { oc =>
-        oc.setDataObject(noc.getDataObject)
-      }
-
-    // Set conditions, target, and PA.
-    editObsComponent(o, SCHEDULING_CONDITIONS, newSiteQualityComponent(EditTargetCC))
-    editObsComponent(o, TELESCOPE_TARGETENV, newTargetComponent(EditTargetName))
+  override def editObservation(o: ISPObservation): Unit = {
+    super.editObservation(o)
     setPositionAngle(o, EditPA)
   }
 
   @Test def testDayCalGetsNeitherConditionsNorTarget(): Unit = {
-    val obsList = instantiate(newObs(DAY_CAL))
+    val obsList = instantiate(newObs(DAY_CAL, Instrument.GmosSouth))
 
     obsList match {
       case o :: Nil =>
@@ -138,7 +105,7 @@ class InstantiationFunctorTest extends SpModelTestBase {
   }
 
   @Test def testReapplyDayCal(): Unit = {
-    val obsList = instantiate(newObs(DAY_CAL))
+    val obsList = instantiate(newObs(DAY_CAL, Instrument.GmosSouth))
     obsList.foreach(editObservation)
 
     reapply(obsList: _*) match {
@@ -146,14 +113,14 @@ class InstantiationFunctorTest extends SpModelTestBase {
         // Everything has been reset.
         assertTrue(siteQuality(o).isEmpty && target(o).isEmpty && getPositionAngle(o) == 0.0)
       case _        =>
-        fail("Exepcted a single observation")
+        fail("Expected a single observation")
     }
 
   }
 
   @Test def testNightCalHasScienceConditions(): Unit = {
     List(ACQ_CAL, PARTNER_CAL, PROG_CAL).foreach { c =>
-      val obsList = instantiate(newObs(c))
+      val obsList = instantiate(newObs(c, Instrument.GmosSouth))
 
       obsList match {
         case o :: Nil =>
@@ -168,7 +135,7 @@ class InstantiationFunctorTest extends SpModelTestBase {
 
   @Test def testScienceHasScienceConditionsAndTarget(): Unit = {
     List(ACQ, SCIENCE).foreach { c =>
-      val obsList = instantiate(newObs(c))
+      val obsList = instantiate(newObs(c, Instrument.GmosSouth))
 
       obsList match {
         case o :: Nil =>
@@ -183,7 +150,7 @@ class InstantiationFunctorTest extends SpModelTestBase {
 
   @Test def testReapplyScienceAndCal(): Unit = {
     List(ACQ, SCIENCE, ACQ_CAL, PARTNER_CAL, PROG_CAL).foreach { c =>
-      val obsList = instantiate(newObs(c))
+      val obsList = instantiate(newObs(c, Instrument.GmosSouth))
       obsList.foreach(editObservation)
 
       reapply(obsList: _*) match {
@@ -197,53 +164,4 @@ class InstantiationFunctorTest extends SpModelTestBase {
       }
     }
   }
-}
-
-object InstantiationFunctorTest {
-    // The program factory interprets a null key as a request to create a new key
-  val WithSomeNewKey: SPNodeKey = null
-
-  val ScienceTargetName = "Biff"
-  val EditTargetName    = "Henderson"
-  val ScienceTargetCC   = CC50
-  val EditTargetCC      = CC80
-  val EditPA            = 42.0
-
-  def newTarget(name: String): SPTarget =
-    new SPTarget() <| (_.setName(name))
-
-  def newTargetObsComp(name: String): TargetObsComp =
-    new TargetObsComp <| (_.setTargetEnvironment(TargetEnvironment.create(newTarget(name))))
-
-  def newSiteQuality(cc: CloudCover): SPSiteQuality =
-    new SPSiteQuality() <| (_.setCloudCover(cc))
-
-  def newParameters(targetName: String, cc: CloudCover): TemplateParameters =
-    TemplateParameters.newInstance(newTarget(targetName), newSiteQuality(cc), new TimeValue(1, TimeValue.Units.hours))
-
-  def siteQuality(o: ISPObservation): Option[SPSiteQuality] =
-    dataObject[SPSiteQuality](o, SCHEDULING_CONDITIONS)
-
-  def target(o: ISPObservation): Option[TargetObsComp] =
-    dataObject[TargetObsComp](o, TELESCOPE_TARGETENV)
-
-  def dataObject[D: Manifest](o: ISPObservation, ct: SPComponentType): Option[D] =
-    obsComp[D](o, ct).map(_._2)
-
-  def obsComp[D: Manifest](o: ISPObservation, ct: SPComponentType): Option[(ISPObsComponent, D)] =
-    o.getObsComponents.asScala.find { _.getType == ct }.map(oc => (oc, oc.getDataObject)).collect {
-      case (oc, d: D) => (oc, d)
-    }
-
-  def getPositionAngle(o: ISPObservation): Double =
-    obsComp[InstGmosSouth](o, INSTRUMENT_GMOSSOUTH).map {
-      case (_, gmos) => gmos.getPosAngleDegrees
-    } | sys.error("expected gmos component")
-
-  def setPositionAngle(o: ISPObservation, pa: Double): Unit =
-    obsComp[InstGmosSouth](o, INSTRUMENT_GMOSSOUTH).foreach {
-      case (oc, gmos) => gmos.setPosAngleDegrees(pa)
-                         oc.setDataObject(gmos)
-      case _          => fail("expected gmos component")
-    }
 }
